proc.tickets, proc.ticks 자료형 변경

+++ kernel/proc.h
@@ -91,8 +91,8 @@ struct proc {
   int killed;                  // If non-zero, have been killed
   int xstate;                  // Exit status to be returned to parent's wait
   int pid;                     // Process ID
-  int tickets;                 // Number of tickets for scheduling
-  int ticks;                   // Number of ticks the process has accumulated
+  uint tickets;                // Number of tickets for scheduling
+  uint64 ticks;                // Number of ticks the process has accumulated
 
   // wait_lock must be held when using this:
   struct proc *parent;         // Parent process
